+++ src/components/BoxContent/style.js
