+++ src/components/Button/index.jsx
@@ -1,0 +1,7 @@
+import { ButtonStyled } from "./style";
+
+function Button({ name, ...rest }) {
+  return <ButtonStyled type="submit">{name}</ButtonStyled>;
+}
+
+export default Button;
