+++ src/components/DivTextResult/style.js
