+++ src/components/SectionResult/style.js
