+++ src/providers/dataProvider/index.jsx
@@ -1,0 +1,15 @@
+import React, { useContext, createContext, useState } from "react";
+
+export const DataContext = createContext();
+
+export const DataProvider = ({ children }) => {
+  const [data, setData] = useState(null);
+
+  return (
+    <DataContext.Provider value={{ data, setData }}>
+      {children}
+    </DataContext.Provider>
+  );
+};
+
+export const useData = () => useContext(DataContext);
